feat: Implement EventController

+++ ewm-service/src/main/java/ru/practicum/yandex/user/dto/EventFullDto.java
@@ -33,7 +33,7 @@ public class EventFullDto {
 
     private String title;
 
-    private long confirmedRequests;
+    private int confirmedRequests;
 
     private long views;
